+++ src/main/kotlin/dev/protsenko/securityLinter/docker/inspection/run/impl/AptIsUsedAnalyzer.kt
@@ -1,0 +1,18 @@
+package dev.protsenko.securityLinter.docker.inspection.run.impl
+
+import com.intellij.codeInspection.ProblemHighlightType
+import com.intellij.codeInspection.ProblemsHolder
+import com.intellij.psi.PsiElement
+import dev.protsenko.securityLinter.core.SecurityPluginBundle
+import dev.protsenko.securityLinter.docker.checker.AptIsUsedValidator
+import dev.protsenko.securityLinter.docker.inspection.run.core.DockerfileRunAnalyzer
+
+class AptIsUsedAnalyzer : DockerfileRunAnalyzer {
+    override fun handle(runCommand: String, psiElement: PsiElement, holder: ProblemsHolder) {
+        if (!AptIsUsedValidator.isValid(runCommand)) {
+            holder.registerProblem(
+                psiElement, SecurityPluginBundle.message("ds032.apt-is-used"), ProblemHighlightType.WEAK_WARNING
+            )
+        }
+    }
+}