+++ src/test/kotlin/dev/protsenko/securityLinter/docker/DS032AptIsUsedInspectionTest.kt
@@ -1,0 +1,11 @@
+package dev.protsenko.securityLinter.docker
+
+import com.intellij.codeInspection.LocalInspectionTool
+import dev.protsenko.securityLinter.core.DockerHighlightingBaseTest
+import dev.protsenko.securityLinter.docker.inspection.run.DockerfileRunInspection
+
+class DS032AptIsUsedInspectionTest(
+    override val ruleFolderName: String = "DS032",
+    override val customFiles: Set<String> = emptySet(),
+    override val targetInspection: LocalInspectionTool = DockerfileRunInspection()
+) : DockerHighlightingBaseTest()